refactor: update movie component organization
